Add liten_up tool to "magically upgrade" Mysa V2 Lite thermostats
Mysa claims that their app is "100% free, really!" but they seem to have
soft-disabled the following features for the Mysa V2 Lite thermostat, their
most inexpensive and compact model 🤨:

1. Zone control

2. Humidity sensor. The Mysa V2 Lite definitely has the sensor and reports its
   value.

   It might be uncalibrated or uncorrected in some way, but it sure looks
   fine to me.

3. Usage graph

   The Mysa V2 Lite *does report* its usage graph to the server,
   including humidity/setpoint/temperature, even though it's normally
   inaccessible to the user.

   Unlike the Mysa V1 or V2 non-lite thermostats, it doesn't have a current
   sensor… just a relay which switches power to the thermostat off/on all
   at all.

   Apparently the device doesn't report any energy usage to the server.

This tool tells the Mysa *server* that your Mysa Lite thermostat (model
BB-V2-0-L) is a Mysa Baseboard V1 thermostat (model BB-V1-1).  This makes
the official apps show the zone control, humidity sensor, and usage graph
options for it.

The Mysa app then sends a slightly-incompatible setpoint signal to the
thermostat, so this tool needs to continue running in order to intercept that
signal and "translate" it to the Mysa V2 Lite format; without it, the app
won't be able to change the device setpoint.

The change is fully and trivially reversible; you can run the script with
the `-R`/`--reset` flag if the Mysa V2 Lite devices get stuck in the wrong
state, or re-pair them if you don't trust the script.

TODO:

- Are there any other messages that need to be translated in order to
  achieve seamless compatibility with the official apps?
- It really should be possible to use the realtime relay on/off signal
  from the Mysa V2 Lite device (sent as `dtyCycle: 0` or `dytCycle: 1` in
  the MQTT packets), along with the estimated current or power of the
  baseboard heater (e.g. 1000 W or 10 A) to publish an *estimate* of the
  energy usage to the server.